refactor: Remove index in features/faceGo, Change import statement
