+++ src/features/faceGo/container/FaceGoContainer.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useContext, useMemo } from "react";
 
-import { FaceGo } from "..";
+import FaceGo from "../presenter/FaceGo";
 import { time } from "../../../common/constants";
 import { game } from "../../../common/property";
 import GameContext from "../../../context/GameContext";
