+++ src/features/faceGo/container/MapContainer.js
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 import useCharacter from "../../../hooks/useCharacter";
 import { map } from "../../../common/constants";
 import { game } from "../../../common/property";
-import { Map } from "../presenter";
+import Map from "../presenter/Map";
 
 const MapContainer = ({
   stage,
