--- src/features/faceGo/container/index.js
@@ -1,5 +1,0 @@
-import FaceRecognitionContainer from "./FaceRecognitionContainer";
-import FaceGoContainer from "./FaceGoContainer";
-import MapContainer from "./MapContainer";
-
-export { FaceRecognitionContainer, FaceGoContainer, MapContainer };
