--- src/features/faceGo/index.js
@@ -1,23 +1,0 @@
-import {
-  FaceRecognitionContainer,
-  FaceGoContainer,
-  MapContainer,
-} from "./container";
-import {
-  FaceDirectionRecord,
-  FaceRecognition,
-  FaceGo,
-  Map,
-  MapCell,
-} from "./presenter";
-
-export {
-  FaceRecognitionContainer,
-  FaceGoContainer,
-  MapContainer,
-  FaceDirectionRecord,
-  FaceRecognition,
-  Map,
-  MapCell,
-  FaceGo,
-};
