--- src/features/faceGo/presenter/index.js
@@ -1,7 +1,0 @@
-import FaceRecognition from "../presenter/FaceRecognition";
-import FaceGo from "../presenter/FaceGo";
-import Map from "../presenter/Map";
-import FaceDirectionRecord from "./FaceDirectionRecord";
-import MapCell from "./MapCell";
-
-export { FaceDirectionRecord, FaceRecognition, FaceGo, Map, MapCell };
