+++ src/features/index.js
@@ -1,24 +1,4 @@
-import {
-  FaceGoContainer,
-  FaceRecognitionContainer,
-  FaceGo,
-  Map,
-  FaceRecognition,
-  FaceDirectionRecord,
-} from "./faceGo";
 import { ShoutGoContainer, ShoutGo } from "./shoutGo";
 import { MainContainer, Main, Page } from "./main";
 
-export {
-  MainContainer,
-  Main,
-  FaceGoContainer,
-  FaceGo,
-  ShoutGoContainer,
-  ShoutGo,
-  Map,
-  FaceRecognitionContainer,
-  FaceRecognition,
-  FaceDirectionRecord,
-  Page,
-};
+export { MainContainer, Main, ShoutGoContainer, ShoutGo, Page };
